[수정] TCA 폴더 분리 ViewModel -> Core
